+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 
 - Brand refresh
 
+## [1.0.13]
+
+### Added
+
+- Added Sentry for error tracking.
+
+### Fixed
+
+- Fixed sidebar menu and notification in mobile view.
+
 ## [1.0.12]
 
 ### Added
@@ -15,6 +25,11 @@
 ### Fixed
 
 - Fetch balance on mount.
+
+## [1.0.12]
+
+- Update starknet packages to latest versions.
+- Fix add network to wallet in rabby wallet.
 
 ## [1.0.11]